Extract circular supplement builder and parser output normalizer
Further slims the NIST builder from 952 to 483 lines by extracting two
more concerns:

- nist/circular_supplement_builder.rb (280 lines): the CIRC/LCIRC
  supplement construction pipeline (was 235 lines inline in builder).
  The three methods (build_circular_supplement, build_circular_supplement_base,
  build_circular_supplement_wrapper) are now in a dedicated class that
  receives the parent Builder for recursive base construction.

- nist/parser_output_normalizer.rb (233 lines): pre-processing logic
  that normalizes parser output before construction. Encapsulates seven
  in-place mutation patterns: edition_e -> update_year, embedded edition
  with/without dash year, second_number_edition_year, FIPS month-year,
  IR compound vs edition, and dash year disambiguation.

Three patterns stay in the builder because they create local variables
that flow into the construction phase (letter_suffix_part,
edition_from_dash_year, edition_from_embedded).
