+++ lib/pubid/nist/builder.rb
@@ -17,6 +17,8 @@ module Pubid
         @scheme = scheme
         @router = Router.new
         @caster = Caster.new
+        @normalizer = ParserOutputNormalizer.new
+        @circular_supplement_builder = CircularSupplementBuilder.new(self)
       end
 
       # Build an identifier object from parsed data
@@ -26,16 +28,10 @@ module Pubid
         # Parslet can return array of hashes - merge them
         parsed_hash = parsed.is_a?(Array) ? flatten_array(parsed) : parsed
 
-        # NEW: Fix for update year captured as edition_e
-        # Pattern: "800-53r4/Upd3-2015" where parser captures "-2015" as :edition_e
-        # but it should be :update_year as part of the update component
-        if parsed_hash[:update_prefix] && parsed_hash[:update] && parsed_hash[:edition_e]
-          # Move edition_e to update_year
-          edition_id = parsed_hash[:edition_e][:edition_id]
-          parsed_hash[:update] =
-            parsed_hash[:update].merge(update_year: edition_id)
-          parsed_hash.delete(:edition_e) # Remove the edition_e hash
-        end
+        # Pure in-place shape corrections: parser-output normalization that
+        # only mutates parsed_hash. Extraction logic that needs to surface
+        # components to the construction phase stays below.
+        @normalizer.normalize(parsed_hash)
 
         # NEW: Fix for letter suffix in number with edition_dash_year
         # Pattern: "304a-2017" where parser returns first_number="304a" and edition_dash_year="2017"
@@ -64,226 +60,6 @@ module Pubid
             edition_from_dash_year = Components::Edition.new(type: "e",
                                                              id: dash_year)
             parsed_hash.delete(:edition_dash_year) # Remove the old key
-          end
-        end
-
-        # NEW: Fix for edition_dash_year with embedded edition in first_number
-        # Pattern: "44e2-1955" where first_number="44e2" and edition_dash_year="1955"
-        # Expected: edition extracted from "44e2" (type: "e", id: "2") with additional_text="1955"
-        #
-        # We need to check for this pattern BEFORE the simple year-as-edition pattern below
-        if parsed_hash[:first_number]&.to_s&.match?(/^[0-9]+[a-zA-Z]\d+$/) && parsed_hash[:edition_dash_year]
-          # Extract edition from embedded pattern (e.g., "44e2" -> type="e", id="2")
-          number_str = parsed_hash[:first_number].to_s
-          if match_data = number_str.match(/^(\d+)([a-zA-Z])(\d+)$/)
-            base_number = match_data[1]
-            edition_type = match_data[2].downcase
-            edition_id = match_data[3]
-
-            # Update first_number to base number only
-            parsed_hash[:first_number] =
-              Components::Code.new(number: base_number)
-
-            # Create Edition with additional_text from edition_dash_year
-            dash_year = parsed_hash[:edition_dash_year][:dash_year]
-            edition_obj = Components::Edition.new(
-              type: edition_type,
-              id: edition_id,
-              additional_text: dash_year,
-            )
-            parsed_hash[:edition_with_year] = edition_obj
-            parsed_hash.delete(:edition_dash_year)
-          end
-        end
-
-        # NEW: Fix for edition embedded in first_number WITHOUT edition_dash_year
-        # Pattern: "8115r1" where parser returns first_number="8115r1" with NO edition_dash_year
-        # Expected: first_number="8115", edition with type="r", id="1"
-        # This handles patterns like IR "8115r1/upd" where r1 is the edition
-        # CRITICAL: Only process if NO second_number is present!
-        # Otherwise, the compound number logic (lines 418-437) will handle edition+year patterns
-        if !parsed_hash[:second_number] &&
-            parsed_hash[:first_number]&.to_s&.match?(/^[0-9]+[a-zA-Z]\d+$/) &&
-            !parsed_hash[:edition_dash_year]
-          number_str = parsed_hash[:first_number].to_s
-          if match_data = number_str.match(/^(\d+)([a-zA-Z])(\d+)$/)
-            base_number = match_data[1]
-            edition_type = match_data[2].downcase
-            edition_id = match_data[3]
-
-            # Update first_number to base number only
-            parsed_hash[:first_number] =
-              Components::Code.new(number: base_number)
-
-            # Create Edition without additional_text (no year in this pattern)
-            edition_obj = Components::Edition.new(
-              type: edition_type,
-              id: edition_id,
-            )
-            parsed_hash[:edition_with_year] = edition_obj
-          end
-        end
-
-        # REMOVED: edition_r_with_space pre-processing
-        # Now handled by cast_edition method with original_prefix preservation
-
-        # NEW: Fix for second_number_edition_year pattern
-        # Pattern: "105-1-1990" where parser returns second_number_edition_year={second_number="1", dash_year="1990"}
-        # Expected: number="105-1", edition with type="e", id="1990", original_prefix="-"
-        if parsed_hash[:second_number_edition_year]
-          second_num_value = parsed_hash[:second_number_edition_year][:second_number]
-          dash_year = parsed_hash[:second_number_edition_year][:dash_year]
-
-          # Extract second_number and edition_dash_year from the combined hash
-          parsed_hash[:second_number] = second_num_value
-
-          # For HB (Handbook) series, create year-only edition with dash rendering
-          is_handbook = begin
-            parsed_hash[:series].to_s == "HB"
-          rescue StandardError
-            false
-          end
-
-          if is_handbook && dash_year.to_s.match?(/^\d{4}$/)
-            # Create Edition with type="e" and id=dash_year for HB series
-            edition_obj = Components::Edition.new(type: "e", id: dash_year)
-            parsed_hash[:edition_from_year] = edition_obj
-          else
-            # For other series, treat dash_year as edition_dash_year for further processing
-            parsed_hash[:edition_dash_year] = { dash_year: dash_year }
-          end
-
-          parsed_hash.delete(:second_number_edition_year)
-        end
-
-        # NEW: Fix for fips_month_year_after_part pattern
-        # Pattern: "11-1-Sep1977" where parser returns fips_month_year_after_part={second_number="1", edition_month="Sep", edition_year="1977"}
-        # Expected: number="11-1", edition with type="e", id="197709" (year + month number)
-        if parsed_hash[:fips_month_year_after_part]
-          second_num_value = parsed_hash[:fips_month_year_after_part][:second_number]
-          month_str = parsed_hash[:fips_month_year_after_part][:edition_month]
-          year_str = parsed_hash[:fips_month_year_after_part][:edition_year]
-
-          # Extract second_number
-          parsed_hash[:second_number] = second_num_value
-
-          # Convert month abbreviation to month number and combine with year
-          month_num = Date::ABBR_MONTHNAMES.index(month_str) ||
-            Date::MONTHNAMES.index(month_str) ||
-            month_str.to_i
-
-          edition_id = if month_num&.positive?
-                         "#{year_str}#{format('%02d', month_num)}"
-                       else
-                         year_str
-                       end
-
-          # Create Edition with type="e" and combined ID
-          edition_obj = Components::Edition.new(type: "e", id: edition_id)
-          parsed_hash[:edition_from_year] = edition_obj
-
-          parsed_hash.delete(:fips_month_year_after_part)
-        end
-
-        # NEW: Fix for IR compound number vs edition pattern
-        # Pattern: "84-2946" where parser returns first_number="84", edition_dash_year={dash_year="2946"}
-        # For IR (InteragencyReport), 4-digit second numbers >= 1000 are typically editions
-        # EXCEPT for:
-        # - Numbers > 2699 which are clearly not valid years (like 2946)
-        # - Patterns with embedded edition_e (like "76-1094e2") which should be compound
-        if parsed_hash[:series]&.to_s == "IR" && parsed_hash[:first_number] && parsed_hash[:edition_dash_year]
-          first_num = parsed_hash[:first_number].to_s
-          dash_year = parsed_hash[:edition_dash_year][:dash_year].to_s
-
-          # Check if first_number looks like a 2-digit year (00-99)
-          if first_num.match?(/^\d{2}$/) && dash_year.match?(/^\d{4}$/)
-            dash_year_num = dash_year.to_i
-            # Valid year range for IR: 1901-2026 (NBS establishment to present)
-            is_valid_year = dash_year_num.between?(1901, 2026)
-            # If there's an embedded edition (e2, e3, etc.), treat as compound, not edition
-            has_embedded_edition = parsed_hash[:edition_e]
-            if is_valid_year && !has_embedded_edition
-              # Edition format: "76e1000", "76e2013", "76e1100", "80e2100", "81e2300", "81e2400", "82e2500", "82e2600"
-              parsed_hash[:first_number] =
-                Components::Code.new(number: first_num)
-              parsed_hash[:edition] =
-                Components::Edition.new(type: "e", id: dash_year)
-            else
-              # Compound number: "84-2946" or "76-1094e2"
-              parsed_hash[:first_number] =
-                Components::Code.new(number: "#{first_num}-#{dash_year}")
-            end
-            parsed_hash.delete(:edition_dash_year)
-          end
-        end
-
-        # NEW: Fix for edition_dash_year that's actually a second_number
-        # Pattern: "250-1039" where parser returns edition_dash_year="1039"
-        # Also handles RPT date ranges: "1946-1947" where edition_dash_year="1947"
-        # but it should be second_number="1039" (not a year - years are 1900+ or 2000+)
-        if parsed_hash[:first_number] && parsed_hash[:edition_dash_year] && !parsed_hash[:first_number].to_s.match?(/^[0-9]+[a-zA-Z]\d+$/)
-          dash_year = parsed_hash[:edition_dash_year][:dash_year].to_s
-          series = begin
-            parsed_hash[:series].to_s
-          rescue StandardError
-            ""
-          end
-
-          # Check if dash_year is a valid year (1901-2026)
-          dash_year_num = dash_year.to_i
-          is_valid_year = dash_year_num.between?(1901, 2026)
-
-          # Check if series uses dash-year as edition (HB, CS, FIPS)
-          # These series convert dash-year to edition ONLY for valid years
-          uses_dash_year_as_edition = ["HB", "CS", "FIPS"].include?(series)
-
-          # GCR always converts dash-year to edition (e.g., "15-1000" -> "15e1000")
-          is_gcr = series == "GCR"
-
-          # IR only converts dash-year to edition for valid years (e.g., "76-1100" -> "76e1100")
-          # but NOT for non-years like 2946 (e.g., "84-2946" stays as "84-2946")
-          is_ir = series == "IR"
-
-          # Check if series uses compound numbers with dash-year (RPT date ranges)
-          is_rpt = series == "RPT"
-
-          if is_rpt
-            # For RPT (Report) series with date ranges, create compound number: "1946-1947"
-            parsed_hash[:first_number] =
-              Components::Code.new(number: "#{parsed_hash[:first_number]}-#{dash_year}")
-            parsed_hash.delete(:edition_dash_year)
-          elsif is_gcr
-            # GCR always converts dash_year to edition regardless of whether it's a valid year
-            # e.g., "15-1000" -> "15e1000", "15-1001" -> "15e1001"
-            edition_obj = Components::Edition.new(type: "e", id: dash_year)
-            parsed_hash[:edition_from_year] = edition_obj
-            parsed_hash.delete(:edition_dash_year)
-          elsif is_ir && is_valid_year
-            # IR only converts valid years to edition
-            # e.g., "76-1100" -> "76e1100", but "84-2946" stays as "84-2946"
-            edition_obj = Components::Edition.new(type: "e", id: dash_year)
-            parsed_hash[:edition_from_year] = edition_obj
-            parsed_hash.delete(:edition_dash_year)
-          elsif uses_dash_year_as_edition && is_valid_year
-            # For HB, CS, FIPS: convert dash_year to edition ONLY if it's a valid year
-            edition_obj = Components::Edition.new(type: "e", id: dash_year)
-            parsed_hash[:edition_from_year] = edition_obj
-            parsed_hash.delete(:edition_dash_year)
-          elsif dash_year.to_i < 1900
-            # For other series with dash_year < 1900, treat as second_number
-            parsed_hash[:second_number] = dash_year
-            parsed_hash.delete(:edition_dash_year)
-          elsif is_valid_year
-            # For remaining cases with valid year (>= 1901), check if series uses it as edition
-            is_handbook = series == "HB"
-            is_commercial_standard = series == "CS"
-            is_fips = series == "FIPS"
-
-            if is_handbook || is_commercial_standard || is_fips
-              edition_obj = Components::Edition.new(type: "e", id: dash_year)
-              parsed_hash[:edition_from_year] = edition_obj
-              parsed_hash.delete(:edition_dash_year)
-            end
           end
         end
 
@@ -697,255 +473,10 @@ module Pubid
         end
       end
 
-      # Build CircularSupplement with base_identifier wrapping
-      # @param parsed_hash [Hash] the parsed supplement data
-      # Build a CIRC/LCIRC supplement. Most forms collapse onto the base
-      # identifier's normal class (Circular / LetterCircular) with isolated
-      # supplement attributes, so they share one model with every other series
-      # and are queryable by part. The V1-compat update forms (slash-year
-      # "118supp3/1926" -> ".../Upd1-192603"; implicit revision "145r11/1925")
-      # render through an Update component the flat supplement model can't
-      # express, so they stay on the CircularSupplement wrapper for now.
+      # Delegate CIRC/LCIRC supplement construction to the dedicated builder.
+      # See CircularSupplementBuilder for the construction pipeline.
       def build_circular_supplement(parsed_hash)
-        if parsed_hash[:supplement_slash_year].is_a?(Hash) ||
-            parsed_hash[:implicit_supplement].is_a?(Hash)
-          return build_circular_supplement_wrapper(parsed_hash)
-        end
-
-        series_value = if parsed_hash[:circ_series].is_a?(Hash)
-                         parsed_hash[:circ_series][:series]
-                       else
-                         parsed_hash[:series]
-                       end
-
-        # Date-range supplement (no base document): a plain base identifier with
-        # no number, carrying the range. parsed_format is left at the default so
-        # a dotted MR input still normalizes to the spaced short form.
-        if parsed_hash[:supplement_date_range].is_a?(Hash)
-          range = parsed_hash[:supplement_date_range]
-          identifier = build({ series: series_value })
-          ms = range[:supp_month_start]&.to_s
-          ys = range[:supp_year_start]&.to_s
-          me = range[:supp_month_end]&.to_s
-          ye = range[:supp_year_end]&.to_s
-          identifier.supplement = supplement_from(
-            value: nil, has_revision: false,
-            range_start: (ms && ys ? "#{ms}#{ys}" : nil),
-            range_end: (me && ye ? "#{me}#{ye}" : nil)
-          )
-          return identifier
-        end
-
-        # Based supplement: build the base, then attach the supplement as an
-        # isolated component on that normal class.
-        identifier = build_circular_supplement_base(parsed_hash, series_value)
-        raw = if parsed_hash[:supplement_month_year]
-                parsed_hash[:supplement_month_year].to_s
-              elsif parsed_hash[:supplement_year]
-                parsed_hash[:supplement_year].to_s
-              else
-                "" # supplement_empty or bare marker
-              end
-        identifier.supplement = Components::Supplement.from_raw(raw)
-        identifier
-      end
-
-      # Build just the base identifier for a based CIRC/LCIRC supplement, from
-      # base_portion (number, optional edition "101e2" or letter suffix "378G")
-      # or the merged first_number fallback. Returns the normal class.
-      def build_circular_supplement_base(parsed_hash, series_value)
-        base_portion = parsed_hash[:base_portion]
-        unless base_portion
-          return build({
-                         publisher: parsed_hash[:publisher],
-                         series: series_value || parsed_hash[:series],
-                         first_number: parsed_hash[:first_number],
-                         parsed_format: parsed_hash[:parsed_format],
-                       })
-        end
-
-        base_number = if base_portion.is_a?(Hash)
-                        base_portion[:simple_number] || base_portion[:base_number]
-                      else
-                        base_portion
-                      end
-
-        letter_suffix = nil
-        if base_portion.is_a?(Hash) && base_portion[:letter_suffix]
-          letter_suffix = base_portion[:letter_suffix].to_s.upcase
-        end
-
-        publisher_value = nil
-        if parsed_hash[:circ_series].is_a?(Hash) && parsed_hash[:circ_series][:series]
-          series_str = parsed_hash[:circ_series][:series].to_s
-          publisher_value = series_str.split.first if series_str.include?(" ")
-        end
-
-        has_edition = base_portion.is_a?(Hash) && base_portion[:edition_number]
-
-        base_number_with_suffix = base_number.to_s
-        base_number_with_suffix += letter_suffix if letter_suffix
-        if has_edition
-          base_number_with_suffix += "e#{base_portion[:edition_number]}"
-        end
-
-        base_hash = {
-          series: series_value,
-          first_number: base_number_with_suffix,
-          parsed_format: parsed_hash[:parsed_format],
-        }
-        base_hash[:publisher] = publisher_value if publisher_value
-        base_hash[:edition_e] =
-          { edition_id: base_portion[:edition_number] } if has_edition
-
-        build(base_hash)
-      end
-
-      # @return [Identifiers::CircularSupplement] the supplement identifier
-      def build_circular_supplement_wrapper(parsed_hash)
-        supplement = Identifiers::CircularSupplement.new
-
-        # Extract series from circ_series if present (nested structure from parser)
-        series_value = nil
-        if parsed_hash[:circ_series].is_a?(Hash)
-          series_value = parsed_hash[:circ_series][:series]
-        elsif parsed_hash[:series]
-          series_value = parsed_hash[:series]
-        end
-
-        # Handle date range supplement (no base)
-        if parsed_hash[:supplement_date_range].is_a?(Hash)
-          range = parsed_hash[:supplement_date_range]
-          month_start = range[:supp_month_start]&.to_s
-          year_start = range[:supp_year_start]&.to_s
-          month_end = range[:supp_month_end]&.to_s
-          year_end = range[:supp_year_end]&.to_s
-
-          supplement.supplement_date_range_start = "#{month_start}#{year_start}" if month_start && year_start
-          supplement.supplement_date_range_end = "#{month_end}#{year_end}" if month_end && year_end
-
-          return supplement
-        end
-
-        # Build base identifier from base_portion (if present)
-        # If not present (because it was merged during parsing), use first_number
-        if parsed_hash[:base_portion]
-          # Extract the actual number value from base_portion hash
-          # base_portion can be: {:simple_number=>"118"}, {:base_number=>"145", :revision_letter=>"r", :revision_number=>"11"}, etc.
-          base_portion = parsed_hash[:base_portion]
-          base_number = if base_portion.is_a?(Hash)
-                          # Extract the value from whichever key is present
-                          base_portion[:simple_number] || base_portion[:base_number]
-                        else
-                          base_portion
-                        end
-
-          # Check for letter suffix in base_portion (e.g., "378G")
-          letter_suffix = nil
-          if base_portion.is_a?(Hash) && base_portion[:letter_suffix]
-            letter_suffix = base_portion[:letter_suffix].to_s.upcase
-          end
-
-          # Extract publisher from circ_series if present
-          publisher_value = nil
-          if parsed_hash[:circ_series].is_a?(Hash) && parsed_hash[:circ_series][:series]
-            series_str = parsed_hash[:circ_series][:series].to_s
-            # Extract publisher from series (e.g., "NBS LCIRC" -> "NBS")
-            publisher_value = series_str.split.first if series_str.include?(" ")
-          end
-
-          # Check if base_portion has revision (for patterns like "145r11/1925")
-          has_revision = base_portion.is_a?(Hash) && base_portion[:revision_letter] && base_portion[:revision_number]
-
-          # NEW: Check if base_portion has edition_number (for patterns like "101e2")
-          has_edition = base_portion.is_a?(Hash) && base_portion[:edition_number]
-
-          # Include letter suffix in base_number if present
-          # Also include edition_number if present (for "101e2" pattern)
-          base_number_with_suffix = base_number.to_s
-          if letter_suffix
-            base_number_with_suffix += letter_suffix
-          end
-          if has_edition
-            base_number_with_suffix += "e#{base_portion[:edition_number]}"
-          end
-
-          # Reconstruct parse hash for base identifier
-          base_hash = {
-            series: series_value,
-            first_number: base_number_with_suffix,
-            parsed_format: parsed_hash[:parsed_format],
-          }
-          base_hash[:publisher] = publisher_value if publisher_value
-
-          # NEW: Add edition_number to base_hash for patterns like "101e2"
-          # This will be processed by the normal build() logic to create Edition component
-          if has_edition
-            # Create edition_e hash that will be converted to Edition with type="e"
-            base_hash[:edition_e] =
-              { edition_id: base_portion[:edition_number] }
-          end
-
-          # Recursively build base identifier
-          # This will go through normal build() process which extracts edition from "101e2"
-          supplement.base_identifier = build(base_hash)
-
-          # NEW: Handle revision + implicit supplement pattern (e.g., "145r11/1925")
-          # Create update format: "Upd1-{year}{revision_number}"
-          if has_revision && parsed_hash[:implicit_supplement].is_a?(Hash)
-            revision_number = base_portion[:revision_number].to_s
-            supplement_year = parsed_hash[:implicit_supplement][:implicit_supplement_year].to_s
-
-            # Create Update component for revision+supplement pattern
-            # Format: Upd1-{year}{revision_number} (always use "1" and concatenate year+revision)
-            update_value = "Upd1-#{supplement_year}#{revision_number}"
-            supplement.update = update_value
-            supplement.implicit_supplement = true # Mark as implicit supplement for rendering
-          end
-        elsif parsed_hash[:first_number]
-          # base_portion was lost during merge, use first_number to build base identifier
-          base_hash = {
-            publisher: parsed_hash[:publisher],
-            series: series_value || parsed_hash[:series],
-            first_number: parsed_hash[:first_number],
-            parsed_format: parsed_hash[:parsed_format],
-          }
-
-          # Recursively build base identifier
-          supplement.base_identifier = build(base_hash)
-        end
-
-        # Build supplement edition from captured data
-        if parsed_hash[:supplement_month_year]
-          # Parse month+year format like "Jan1924"
-          month_year = parsed_hash[:supplement_month_year].to_s
-          supplement.edition = Components::Edition.new(type: "s",
-                                                       id: month_year)
-        elsif parsed_hash[:supplement_year]
-          # Just year: 1924
-          supplement.edition = Components::Edition.new(type: "s",
-                                                       id: parsed_hash[:supplement_year].to_s)
-        elsif parsed_hash[:supplement_slash_year].is_a?(Hash)
-          # NEW: Handle supplement_slash_year pattern (e.g., "sup12/1926", "sup1/1927")
-          # V1 format: "Upd1-192612" where "1" is fixed and "192612" is year+number concatenated
-          # Single digit numbers are zero-padded: "sup1/1927" -> "Upd1-192701"
-          supp_hash = parsed_hash[:supplement_slash_year]
-          supp_number = supp_hash[:supp_number]&.to_s
-          supp_year = supp_hash[:supp_year]&.to_s
-
-          # Pad supplement number to 2 digits for single-digit numbers
-          supp_number_padded = supp_number.rjust(2, "0")
-
-          # Create Update component for supplement (V1 compatibility uses Update for supplements)
-          # Format: Upd1-{year}{padded_number} (always use "1" and concatenate year+padded_number)
-          update_value = "Upd1-#{supp_year}#{supp_number_padded}"
-          supplement.update = update_value
-        elsif parsed_hash[:supplement_empty]
-          # Empty supplement - no edition
-          # supplement.edition remains nil
-        end
-
-        supplement
+        @circular_supplement_builder.build_circular_supplement(parsed_hash)
       end
     end
   end
